// applets/radarr/api/model/movieStatusType.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Radarr
 * Radarr API docs
 * OpenAPI spec version: 3.0.0
 */

export type MovieStatusType =
  (typeof MovieStatusType)[keyof typeof MovieStatusType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MovieStatusType = {
  tba: "tba",
  announced: "announced",
  inCinemas: "inCinemas",
  released: "released",
  deleted: "deleted",
} as const;